\name{bdendo11}
\alias{bdendo11}
\docType{data}
\title{A 1:1 subset of the endometrial cancer case-control study}
\description{
The \code{bdendo11} data frame has 126 rows and 13 columns.
This is a subset of the dataset \code{\link{bdendo}} in which each case
was matched with a single control.
}
\source{
Breslow NE, and Day N, Statistical Methods in Cancer Research. Volume
I: The Analysis of Case-Control Studies. IARC Scientific
Publications, IARC:Lyon, 1980.  
}
\examples{
data(bdendo11)
}
\keyword{datasets}
