\name{blcaIT}
\alias{blcaIT}
\docType{data}
\title{Bladder cancer mortality in Italian males}
\description{
Number of deaths from bladder cancer and person-years in the Italian
male population 1955--1979, in ages 25--79.
}
% \usage{data(blcaIT)}
\format{
A data frame with 55 observations on the following 4 variables:
\tabular{rl}{
  \code{age}:    \tab Age at death. Left endpoint of age class \cr
  \code{period}: \tab Period of death. Left endpoint of period \cr
  \code{D}:      \tab Number of deaths \cr
  \code{Y}:      \tab Number of person-years.
}
}
% \source{
%   Reference to a source...
% }
% \references{
%   Reference to a publication...
% }
\examples{
data(blcaIT)
}
\keyword{datasets}
