\docType{package}
\name{Epi}
\alias{Epi}
\alias{Epi-package}
\title{Epi: Functions for manipulation and statistical analysis of epidemiological data}
\description{
\pkg{Epi} has grown out of the course 'Statistical Practise in
Epidemiology with R' \url{http://bendixcarstensen.com/SPE/}.

The major contributions from this course have been the
\code{\link{stat.table}} function for advanced tabulation and summary,
and the functions for representation and the \code{\link{Lexis}}
function(s) for manipulation of multistate data with multiple time
scales.
} 
\details{Click on the \code{Index} link below the line to access
  vignettes (tutorial documents) and an alphabetic list of the functions
in \code{Epi}.}